third_party/blink/web_tests/platform/generic/svg/dom/method-argument-aritychecks-expected.txt

Check that incorrect number of arguments throw TypeError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

SVGAnimationElement


beginElementAt(float offset)
PASS animateElm.beginElementAt() threw exception TypeError: Failed to execute 'beginElementAt' on 'SVGAnimationElement': 1 argument required, but only 0 present..
PASS animateElm.beginElementAt(0) did not throw exception.


endElementAt(float offset)
PASS animateElm.endElementAt() threw exception TypeError: Failed to execute 'endElementAt' on 'SVGAnimationElement': 1 argument required, but only 0 present..
PASS animateElm.endElementAt(0) did not throw exception.




SVGFEDropShadowElement


setStdDeviation(float stdDeviationX, float stdDeviationY)
PASS dropShadow.setStdDeviation() threw exception TypeError: Failed to execute 'setStdDeviation' on 'SVGFEDropShadowElement': 2 arguments required, but only 0 present..
PASS dropShadow.setStdDeviation(0) threw exception TypeError: Failed to execute 'setStdDeviation' on 'SVGFEDropShadowElement': 2 arguments required, but only 1 present..
PASS dropShadow.setStdDeviation(0, 0) did not throw exception.




SVGFEGaussianBlurElement


setStdDeviation(float stdDeviationX, float stdDeviationY)
PASS gaussian.setStdDeviation() threw exception TypeError: Failed to execute 'setStdDeviation' on 'SVGFEGaussianBlurElement': 2 arguments required, but only 0 present..
PASS gaussian.setStdDeviation(0) threw exception TypeError: Failed to execute 'setStdDeviation' on 'SVGFEGaussianBlurElement': 2 arguments required, but only 1 present..
PASS gaussian.setStdDeviation(0, 0) did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
